[REFACTOR] : Data 내부객체에서 별도의 DTO로 분리
